+++ eeUtil/eeutil.py
@@ -14,14 +14,17 @@ GEE_SERVICE_ACCOUNT = os.getenv("GEE_SERVICE_ACCOUNT") or "service account"
 GOOGLE_APPLICATION_CREDENTIALS = os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
 GEE_PROJECT = os.getenv("GEE_PROJECT") or os.getenv("CLOUDSDK_CORE_PROJECT")
 GEE_STAGING_BUCKET = os.getenv("GEE_STAGING_BUCKET")
+GEE_STAGING_BUCKET_PREFIX = os.getenv("GEE_STAGING_BUCKET_PREFIX")
 
 # Unary GEE home directory
-_home = ''
+_cwd = ''
+_gs_bucket_prefix = ''
 
 
 def init(service_account=GEE_SERVICE_ACCOUNT,
          credential_path=GOOGLE_APPLICATION_CREDENTIALS,
          project=GEE_PROJECT, bucket=GEE_STAGING_BUCKET,
+         bucket_prefix=GEE_STAGING_BUCKET_PREFIX, 
          credential_json=GEE_JSON):
     '''
     Initialize Earth Engine and Google Storage bucket connection.
@@ -37,25 +40,30 @@ def init(service_account=GEE_SERVICE_ACCOUNT,
     `credential_path` Path to json file containing private key
     `project`         GCP project for earthengine and storage bucket
     `bucket`          Storage bucket for staging assets for ingestion
+    `bucket_prefix`   Default bucket folder for staging operations
     `credential_json` Json-string to use instead of `credential_path`
 
     https://developers.google.com/earth-engine/service_account
     '''
+    global _gs_bucket_prefix
     init_opts = {}
     if credential_json:
         init_opts['credentials'] = ee.ServiceAccountCredentials(service_account, key_data=credential_json)
     elif credential_path:
         init_opts['credentials'] = ee.ServiceAccountCredentials(service_account, key_file=credential_path)
+        os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = credential_path
     if project:
         init_opts['project'] = project
     ee.Initialize(**init_opts)
-    if credential_path:
-        os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = credential_path
+    if bucket_prefix:
+        init_opts['prefix'] = bucket_prefix
     try:
         gsbucket.init(bucket, **init_opts)
     except Exception as e:
         logging.warning("Could not initialize Google Cloud Storage Bucket.")
         logging.error(e)
+    if bucket_prefix:
+        _gs_bucket_prefix = bucket_prefix
 
 
 def initJson(credential_json=GEE_JSON, project=GEE_PROJECT,
@@ -70,26 +78,37 @@ def initJson(credential_json=GEE_JSON, project=GEE_PROJECT,
 
 def getHome():
     '''Get user root directory'''
-    global _home
     assetRoots = ee.data.getAssetRoots()
     project = ee._cloud_api_utils._cloud_api_user_project
-    if project == 'earthengine-legacy':
+    if project == ee.data.DEFAULT_CLOUD_API_USER_PROJECT:
         if not len(assetRoots):
             raise Exception(f"No available assets for provided credentials in project {project}")
-        _home = assetRoots[0]['id']
+        return assetRoots[0]['id']
     else:
-        _home = f'projects/{project}/assets/'
-    return _home
+        return f'projects/{project}/assets/'
 
 
-def _getHome():
-    '''Cached get user root directory'''
-    global _home
-    return _home if _home else getHome()
+def getCWD():
+    '''Get current directory or root directory'''
+    global _cwd
+    if not _cwd:
+        _cwd = getHome()
+    return _cwd
+
+
+def cd(path):
+    '''Change CWD'''
+    global _cwd 
+    path = os.path.normpath(_path(path))
+    if isFolder(path):
+        _cwd = path
+    else:
+        raise Exception(f"{path} is not a folder")
+    return _cwd
 
 
 def _path(path):
-    '''Add asset root directory to path if not already existing'''
+    '''Add cwd to path if not full path'''
     if path:
         if path[0] == '/':
             return path[1:]
@@ -98,13 +117,13 @@ def _path(path):
         elif len(path) > 9 and path[:9] == 'projects/':
             return path
         else:
-            return os.path.join(_getHome(), path)
-    return _getHome()
+            return os.path.join(getCWD(), path)
+    return getCWD()
 
 
 def getQuota():
     '''Get GEE usage quota'''
-    return ee.data.getAssetRootQuota(_getHome())
+    return ee.data.getAssetRootQuota(getHome())
 
 
 def info(asset=''):
@@ -115,6 +134,17 @@ def info(asset=''):
 def exists(asset):
     '''Check if asset exists'''
     return True if info(asset) else False
+
+
+def isFolder(asset):
+    '''Check if path is folder or imageCollection'''
+    if ee._cloud_api_utils.is_asset_root(asset):
+        return True
+    asset_info = info(asset)
+    return asset_info and asset_info['type'] in (ee.data.ASSET_TYPE_FOLDER, 
+                                                 ee.data.ASSET_TYPE_FOLDER_CLOUD,
+                                                 ee.data.ASSET_TYPE_IMAGE_COLL,
+                                                 ee.data.ASSET_TYPE_IMAGE_COLL_CLOUD)
 
 
 def ls(path='', abspath=False):
@@ -156,19 +186,80 @@ def setProperties(asset, properties={}):
     return ee.data.setAssetProperties(_path(asset), properties)
 
 
-def createFolder(path, imageCollection=False, overwrite=False,
+def createFolder(path, image_collection=False, overwrite=False,
                  public=False):
-    '''Create folder or image collection'''
-    ftype = (ee.data.ASSET_TYPE_IMAGE_COLL if imageCollection
-             else ee.data.ASSET_TYPE_FOLDER)
-    ee.data.createAsset({'type': ftype}, _path(path), overwrite)
+    '''Create folder or image collection, 
+    
+    Automatically creates intermediate folders a la `mkdir -p`
+    '''
+    path = _path(path)
+    upper = os.path.split(path)[0]
+    if not isFolder(upper):
+        createFolder(upper)
+    if overwrite or not isFolder(path):
+        ftype = (ee.data.ASSET_TYPE_IMAGE_COLL if imageCollection
+                 else ee.data.ASSET_TYPE_FOLDER)
+        logging.debug(f'Created {ftype} {path}')
+        ee.data.createAsset({'type': ftype}, path, overwrite)
     if public:
-        setAcl(_path(path), 'public')
+        setAcl(path, 'public')
 
 
 def createImageCollection(path, overwrite=False, public=False):
     '''Create image collection'''
     createFolder(path, True, overwrite, public)
+
+
+def copy(src, dest, overwrite=False, recursive=False):
+    '''Copy asset'''
+    if recursive and isFolder(src):
+        children = ls(src)
+        is_image_collection = info(src)['type'] in (ee.data.ASSET_TYPE_IMAGE_COLL_CLOUD,
+                                                    ee.data.ASSET_TYPE_IMAGE_COLL)
+        createFolder(dest, is_image_collection)
+        for child in children:
+            copy(os.path.join(src, child), os.path.join(dest, child), overwrite, recursive)
+    else:
+        ee.data.copyAsset(_path(src), _path(dest), overwrite)
+
+
+def move(src, dest, overwrite=False, recursive=False):
+    '''Move asset'''
+    src = _path(src)
+    copy(src, _path(dest), overwrite, recursive=False)
+    removeAsset(src, recursive)
+
+
+def removeAsset(asset, recursive=False):
+    '''Delete asset from GEE'''
+    if recursive:
+        if isFolder(asset):
+            for child in ls(asset, abspath=True):
+                removeAsset(child)
+    logging.debug('Deleting asset {}'.format(asset))
+    ee.data.deleteAsset(_path(asset))
+
+#alias
+remove = removeAsset
+
+
+def formatDate(date):
+    '''Format date as ms since last epoch'''
+    if isinstance(date, int):
+        return date
+    seconds = (date - datetime.datetime.utcfromtimestamp(0)).total_seconds()
+    return int(seconds * 1000)
+
+
+def setBucketPrefix(prefix=''):
+    '''Set the default prefix to be used for storage bucket operations'''
+    global _gs_bucket_prefix
+    _gs_bucket_prefix = prefix
+
+
+def getTasks():
+    '''Return a list of all recent tasks'''
+    return ee.data.getTaskList()
 
 
 def _checkTaskCompleted(task_id):
@@ -187,8 +278,18 @@ def _checkTaskCompleted(task_id):
     return False
 
 
-def waitForTasks(task_ids, timeout=3600):
-    '''Wait for tasks to complete, fail, or timeout'''
+def waitForTasks(task_ids=[], timeout=3600):
+    '''Wait for tasks to complete, fail, or timeout
+    
+    Waits for all active tasks if task_ids is not provided
+    '''
+    if not task_ids:
+        task_ids = [t['id'] for t in getTasks() if t['state'] in (
+            ee.batch.Task.State.READY,
+            ee.batch.Task.State.RUNNING,
+            ee.batch.Task.State.UNSUBMITTED,
+        )]
+
     start = time.time()
     elapsed = 0
     while elapsed < timeout:
@@ -207,53 +308,6 @@ def waitForTasks(task_ids, timeout=3600):
 def waitForTask(task_id, timeout=3600):
     '''Wait for task to complete, fail, or timeout'''
     return waitForTasks([task_id], timeout)
-
-
-def getTasks():
-    '''Return a list of all recent tasks'''
-    return ee.data.getTaskList()
-
-
-def copy(src, dest, allowOverwrite=False, recursive=False):
-    '''Copy asset'''
-    src_type = info(src)['type']
-    if recursive and src_type in (ee.data.ASSET_TYPE_FOLDER,
-                                  ee.data.ASSET_TYPE_IMAGE_COLL):
-        children = ls(src)
-        createFolder(dest, (src_type == ee.data.ASSET_TYPE_IMAGE_COLL))
-        for child in children:
-            copy(os.path.join(src, child), os.path.join(dest, child), allowOverwrite, recursive)
-    else:
-        ee.data.copyAsset(_path(src), _path(dest), allowOverwrite)
-
-
-def move(src, dest, allowOverwrite=False, recursive=False):
-    '''Move asset'''
-    src = _path(src)
-    copy(src, _path(dest), allowOverwrite, recursive=False)
-    removeAsset(src, recursive)
-
-
-def removeAsset(asset, recursive=False):
-    '''Delete asset from GEE'''
-    if recursive:
-        if info(asset)['type'] in (ee.data.ASSET_TYPE_FOLDER,
-                                   ee.data.ASSET_TYPE_IMAGE_COLL):
-            for child in ls(asset, abspath=True):
-                removeAsset(child)
-    logging.debug('Deleting asset {}'.format(asset))
-    ee.data.deleteAsset(_path(asset))
-
-#alias
-remove = removeAsset
-
-
-def formatDate(date):
-    '''Format date as ms since last epoch'''
-    if isinstance(date, int):
-        return date
-    seconds = (date - datetime.datetime.utcfromtimestamp(0)).total_seconds()
-    return int(seconds * 1000)
 
 
 def ingestAsset(gs_uri, asset, date='', wait_timeout=0, bands=[]):
@@ -276,12 +330,12 @@ def ingestAsset(gs_uri, asset, date='', wait_timeout=0, bands=[]):
             bands = [{'id': b} for b in bands]
         params['bands'] = bands
     request_id = ee.data.newTaskId()[0]
-    task = ee.data.startIngestion(request_id, params, True)
-    logging.debug(f"Ingesting {gs_uri} to {asset}: {task['id']}")
+    task_id = ee.data.startIngestion(request_id, params, True)['id']
+    logging.debug(f"Ingesting {gs_uri} to {asset}: {task_id}")
     if wait_timeout:
         waitForTask(task_id, wait_timeout)
     
-    return task['id']
+    return task_id
 
 
 def uploadAsset(filename, asset, gs_prefix='', date='', public=False,
@@ -292,25 +346,15 @@ def uploadAsset(filename, asset, gs_prefix='', date='', public=False,
     `file`         local file paths
     `asset`        destination path
     `gs_prefix`    GS folder for staging (else files are staged to bucket root)
-    `date`         Optional date tag (datetime.datetime or int ms since epoch)
+    `date`         optional date tag (datetime.datetime or int ms since epoch)
     `public`       set acl public if True
     `timeout`      wait timeout secs for completion of GEE ingestion
     `clean`        delete files from GS after completion
+    `bands`        band names to assign
     '''
-    gs_uris = gsbucket.stage(filename, gs_prefix)
-    try:
-        ingestAsset(gs_uris[0], asset, date, timeout, bands)
-        if public:
-            setAcl(asset, 'public')
-    except Exception as e:
-        logging.error(e)
-    if clean:
-        gsbucket.remove(gs_uris)
-    
-    return asset
+    dates = [date] if date else []
+    return uploadAssets([filename], [asset], gs_prefix, dates, public, timeout, clean, bands)[0]
 
-# ALIAS
-upload = uploadAsset
 
 def uploadAssets(files, assets, gs_prefix='', dates=[], public=False,
                  timeout=3600, clean=True, bands=[]):
@@ -320,22 +364,24 @@ def uploadAssets(files, assets, gs_prefix='', dates=[], public=False,
     `files`        local file paths
     `assets`       destination paths
     `gs_prefix`    GS folder for staging (else files are staged to bucket root)
-    `dates`        Optional date tags (datetime.datetime or int ms since epoch)
+    `dates`        optional date tags (datetime.datetime or int ms since epoch)
     `public`       set acl public if True
     `timeout`      wait timeout secs for completion of GEE ingestion
     `clean`        delete files from GS after completion
+    `bands`        band names to assign, all assets must have the same number of bands
     '''
+    gs_prefix = gs_prefix or _gs_bucket_prefix
+    task_ids = []
     if len(assets) != len(files):
         raise Exception(f"Files and assets must be of same length. Found {len(files)}, {len(assets)}")
     if len(dates) and len(dates) != len(assets):
         raise Exception(f"Must have one date per asset if dates specified. Found {len(dates)}, {len(assets)}")
-
-    gs_uris = gsbucket.stage(files, gs_prefix)
     if not dates:
         dates = ['' for i in range(len(assets))]
     
-    task_ids = [ingestAsset(gs_uris[i], assets[i], dates[i], 0, bands)
-                for i in range(len(files))]
+    gs_uris = gsbucket.stage(files, gs_prefix)
+    for i in range(len(files)):
+        task_ids.append(ingestAsset(gs_uris[i], assets[i], dates[i], 0, bands))
     
     try:
         waitForTasks(task_ids, timeout)
@@ -344,70 +390,45 @@ def uploadAssets(files, assets, gs_prefix='', dates=[], public=False,
                 setAcl(asset, 'public')
     except Exception as e:
         logging.error(e)
-    
     if clean:
         gsbucket.remove(gs_uris)
     
     return assets
 
+# ALIAS
+upload = uploadAssets
 
-def downloadAsset(asset, filename=None, gs_prefix='', timeout=3600, clean=True, **kwargs):
+
+def downloadAsset(asset, directory=None, gs_prefix='', timeout=3600, clean=True, **kwargs):
     '''Export image asset to GS and download to local machine
 
     `asset`     Asset ID
-    `filename`  Optional filename for export otherwise defaults to Asset ID
+    `directory` Optional local directory to save assets to
     `gs_prefix` GS folder for staging (else files are staged to bucket root)
-    `timeout`   Wait timeout secs for export task completion
+    `timeout`   Wait timeout secs for GEE export task completion
     `clean`     Remove file from GS after download
     `kwargs`    Additional args to pass to ee.batch.Export.image.toCloudStorage()
     '''
-    if filename is None:
-        filename = os.path.basename(asset)
-    else:
-        # .tif is automatically appended...
-        filename = os.path.splitext(filename)[0]
-    
-    image = ee.Image(asset)
-    path = os.path.join(gs_prefix, filename)
-    task = ee.batch.Export.image.toCloudStorage(
-        image,
-        bucket=gsbucket.getName(),
-        fileNamePrefix=path,
-        **kwargs
-    )
-    task.start()
-
-    uri = f"{gsbucket.asURI(path)}.tif"
-    logging.debug(f"Exporting asset {asset} to {uri}")
-
-    try:
-        waitForTask(task.id, timeout)
-        gsbucket.download(uri, filename)
-        if clean:
-            gsbucket.remove(uri)
-    except Exception as e:
-        logging.error(e)
-    
-    return filename
+    return downloadAssets([asset], directory, gs_prefix, clean, timeout, **kwargs)
 
 
-# ALIAS
-download = downloadAsset
-
-
-def downloadAssets(assets, gs_prefix='', clean=True, timeout=3600, **kwargs):
+def downloadAssets(assets, directory=None, gs_prefix='', clean=True, timeout=3600, **kwargs):
     '''Export image assets to GS and download to local machine
 
     `asset`     Asset ID
+    `directory` Optional local directory to save assets to
     `gs_prefix` GS folder for staging (else files are staged to bucket root)
-    `timeout`   Wait timeout secs for export task completion
+    `timeout`   Wait timeout secs for GEE export task completion
     `clean`     Remove file from GS after download
     `kwargs`    Additional args to pass to ee.batch.Export.image.toCloudStorage()
     '''
+    gs_prefix = gs_prefix or _gs_bucket_prefix
     task_ids = []
     uris = []
+    if not os.path.isdir(directory):
+        raise Exception(f"Folder {directory} does not exist")
     for asset in assets:
-        image = ee.Image(asset)
+        image = ee.Image(_path(asset))
         path = os.path.join(gs_prefix, os.path.basename(asset))
 
         task = ee.batch.Export.image.toCloudStorage(
@@ -429,10 +450,13 @@ def downloadAssets(assets, gs_prefix='', clean=True, timeout=3600, **kwargs):
     try:
         waitForTasks(task_ids, timeout)
         for uri in uris:
-            gsbucket.download(uri)
+            gsbucket.download(uri, directory=directory)
             if clean:
                 gsbucket.remove(uri)
     except Exception as e:
         logging.error(e)
 
     return filenames
+
+# ALIAS
+download = downloadAssets