+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG
 
-## v0.2.0 - 2020-03-15
+## v0.2.0 - 2020-03-21
 ### added
-- add recursive copy, move, remove operations
-- downloadAsset function
+- added recursive copy, move, remove operations.
+- added `download` function to export and download assets via GCS.
+- `createFolder` now behaves like `mkdir -p`.
+- added `getCWD`, `cd` to get/set working directory.
+- added optional `bucket_prefix` parameter to `init()` to set default bucket
+folder for staging operations. 
 
 ### changed
 - Compatible with earthengine cloud api